+++ CHANGELOG.md
@@ -10,10 +10,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Examples: Add Relay JavaScript example.
+- **Breaking**: Moved utf-8 conversion functions to `utils`.
 
 ### Fixed
 
-- Replace Base 64 buggy conversion functions with `uint8arrays`. 
+- Replace Base 64 buggy conversion functions with `uint8arrays`.
+
+### Removed
+
+- **Breaking**: Removed `equalByteArrays`, use `uint8arrays/equals` instead.
+  See changes in `eth-pm` example.
+- **Breaking**: Removed deprecated utils functions.
 
 ## [0.19.2] - 2022-03-21